generate_samplesheet: add --group_across_illumina_lanes argument
We are often working with Illumina data, and a very common way to run a full
flowcell is to prepare a single library that you run across all the lanes.  For
example, this is how all the data the NAO is receiving from Marc Johnson's lab
looks.  With this data the _Lnnn suffix on each sample tells you what lane the
data came from, which isn't meaningful.  Add an option to group across lanes,
so samples 'foo_L001' and 'foo_L002' are the same group, 'foo'.

Also add -e and -u so the script will exit if a command fails or tries to read
an undefined variable. This is best practice for bash scripts.

Separately, this script is too complex to be in bash, and should be python.