Swap normal run workflow for lca-based run workflows
This PR removes the LCA run workflows for the normal run workflows,
getting rid of the old one as we will no longer do this anymore. This
includes functionality for both the short and long read pipeline. The
functionality for downstream and docs will come in a later PR (or
previous depending on when this is committed).