FILTER_VIRAL_SAM CP alignment bug fix (#414)
Fixed FILTER_VIRAL_SAM grouping bug for concordant pairs with identical positions but different alignment scores. This bug resulted in a sorting issue that lead to two forward reads being consecutive in the SAM file, breaking the requirement of PROCESS_VIRAL_BOWTIE2_SAM.

The way FILTER_VIRAL_SAM works is that we try to group reads with their forward/reverse mate if they have one. In the case of paired reads, we try to find a unique key. In this case, the unique key was not unique, resulting in 2 separate secondary alignments being grouped together. Once in a group, we can usually just sort by flag to get the forward and reverse reads in the order required by PROCESS_VIRAL_BOWTIE2_SAM; however, with 2 separate secondary alignments, this would cause the forward alignments to show up consecutively. This change updates the key used for grouping such that it is unique.