Allow DOWNSTREAM to run even if samples have no viral hits. (#415)
Solves issue #280 by adding a module that removes samples from the grouping file if they have no viral hits (and reports them in a separate file). We also maintain the previous behavior, which was throw an error if a sample is in the viral hits table, but not in the grouping file.